rename "parseComponent"
