+++ package/src/scripts/helpers/parseComponent.ts
